updated measurements migration

+++ priv/repo/migrations/20250407222554_create_measurements.exs
@@ -9,8 +9,8 @@ defmodule PoolProj.Repo.Migrations.CreateMeasurements do
       add :free_chlorine, :float
       add :combined_chlorine, :float
       add :pH, :float
-      add :total_alkalinity, :integer
-      add :calcium_hardness, :integer
+      add :total_alkalinity, :float
+      add :calcium_hardness, :float
 
       timestamps()
     end
@@ -18,4 +18,3 @@ defmodule PoolProj.Repo.Migrations.CreateMeasurements do
     create index(:measurements, [:pool_id])
   end
 end
-
